LabFolder_Generator/requirements.txt
pip~=25.0.1
python-docx~=1.1.2
PyQt6>=6.6.0
google-generativeai>=0.8.5
PyPDF2>=3.0.0
python-docx>=0.8.11
stdlib-list>=0.11.1